+++ src/main/java/com/itp4511/domain/Guestlist.java
@@ -2,16 +2,17 @@ package com.itp4511.domain;
 
 
 import java.io.Serializable;
+import java.util.ArrayList;
 
 public class Guestlist implements Serializable {
 
-  private long guestListId;
+  private int guestListId;
 
 
   public Guestlist() {
   }
 
-  public Guestlist(long guestListId ) {
+  public Guestlist(int guestListId ) {
     this.guestListId = guestListId;
 
   }
@@ -20,9 +21,48 @@ public class Guestlist implements Serializable {
     return guestListId;
   }
 
-  public void setGuestListId(long guestListId) {
+  public void setGuestListId(int guestListId) {
     this.guestListId = guestListId;
   }
+
+
+
+
+
+
+
+
+//
+//
+//  //customer method  替代json
+//
+//
+//  private ArrayList<Guest> guest;
+//  // getter and setter of customers.
+//  // default constructor
+//
+//  public Guestlist(Guest guest){
+//    guest = new ArrayList<Guest>();
+//  }
+//
+//
+//  public ArrayList<CustomerBean> getCustomers() {
+//    return customers;
+//  }
+//
+//
+//  public void setCustomers(ArrayList s){
+//    customers = s;
+//  }
+//
+//  public void addCustomer(CustomerBean customer) {
+//    customers.add(customer);
+//  }
+//
+//
+
+
+
